Add testing utility to check for traitsui.
